Reconcile the two laplace approximation functions
