+++ tests/test_jax_find_map.py
@@ -5,7 +5,6 @@ import pytest
 
 from pymc_experimental.inference.find_map import (
     find_MAP,
-    fit_laplace,
     scipy_optimize_funcs_from_loss,
 )
 
@@ -96,113 +95,3 @@ def test_JAX_map(method, use_grad, use_hess, use_jax_gradients, rng):
 
     assert np.isclose(mu_hat, 3, atol=0.5)
     assert np.isclose(np.exp(log_sigma_hat), 1.5, atol=0.5)
-
-
-@pytest.mark.parametrize(
-    "transform_samples",
-    [True, False],
-    ids=["transformed", "untransformed"],
-)
-@pytest.mark.parametrize("mode", ["JAX", None], ids=["jax", "pytensor"])
-def test_fit_laplace_coords(rng, transform_samples, mode):
-    coords = {"city": ["A", "B", "C"], "obs_idx": np.arange(100)}
-    with pm.Model(coords=coords) as model:
-        mu = pm.Normal("mu", mu=3, sigma=0.5, dims=["city"])
-        sigma = pm.Normal("sigma", mu=1.5, sigma=0.5, dims=["city"])
-        obs = pm.Normal(
-            "obs",
-            mu=mu,
-            sigma=sigma,
-            observed=rng.normal(loc=3, scale=1.5, size=(100, 3)),
-            dims=["obs_idx", "city"],
-        )
-
-        optimized_point = find_MAP(
-            method="trust-ncg",
-            use_grad=True,
-            use_hessp=True,
-            progressbar=False,
-            compile_kwargs=dict(mode=mode),
-            use_jax_gradients=mode == "JAX",
-        )
-
-    for value in optimized_point.values():
-        assert value.shape == (3,)
-
-    idata = fit_laplace(
-        optimized_point,
-        model,
-        transform_samples=transform_samples,
-        progressbar=False,
-        compile_kwargs=dict(mode=mode),
-    )
-
-    np.testing.assert_allclose(np.mean(idata.posterior.mu, axis=1), np.full((2, 3), 3), atol=0.5)
-    np.testing.assert_allclose(
-        np.mean(idata.posterior.sigma, axis=1), np.full((2, 3), 1.5), atol=0.3
-    )
-
-
-def test_fit_laplace_ragged_coords(rng):
-    coords = {"city": ["A", "B", "C"], "feature": [0, 1], "obs_idx": np.arange(100)}
-    with pm.Model(coords=coords) as ragged_dim_model:
-        X = pm.Data("X", np.ones((100, 2)), dims=["obs_idx", "feature"])
-        beta = pm.Normal(
-            "beta", mu=[[-100.0, 100.0], [-100.0, 100.0], [-100.0, 100.0]], dims=["city", "feature"]
-        )
-        mu = pm.Deterministic(
-            "mu", (X[:, None, :] * beta[None]).sum(axis=-1), dims=["obs_idx", "feature"]
-        )
-        sigma = pm.Normal("sigma", mu=1.5, sigma=0.5, dims=["city"])
-
-        obs = pm.Normal(
-            "obs",
-            mu=mu,
-            sigma=sigma,
-            observed=rng.normal(loc=3, scale=1.5, size=(100, 3)),
-            dims=["obs_idx", "city"],
-        )
-
-        optimized_point, _ = find_MAP(
-            method="Newton-CG", use_grad=True, progressbar=False, return_raw=True
-        )
-
-    idata = fit_laplace(optimized_point, ragged_dim_model, progressbar=False)
-
-    assert idata["posterior"].beta.shape[-2:] == (3, 2)
-    assert idata["posterior"].sigma.shape[-1:] == (3,)
-
-    # Check that everything got unraveled correctly -- feature 0 should be strictly negative, feature 1
-    # strictly positive
-    assert (idata["posterior"].beta.sel(feature=0).to_numpy() < 0).all()
-    assert (idata["posterior"].beta.sel(feature=1).to_numpy() > 0).all()
-
-
-@pytest.mark.parametrize(
-    "transform_samples",
-    [True, False],
-    ids=["transformed", "untransformed"],
-)
-def test_fit_laplace(transform_samples):
-    with pm.Model() as simp_model:
-        mu = pm.Normal("mu", mu=3, sigma=0.5)
-        sigma = pm.Normal("sigma", mu=1.5, sigma=0.5)
-        obs = pm.Normal(
-            "obs",
-            mu=mu,
-            sigma=sigma,
-            observed=np.random.default_rng().normal(loc=3, scale=1.5, size=(10000,)),
-        )
-
-        optimized_point = find_MAP(
-            method="Newton-CG",
-            use_grad=True,
-            progressbar=False,
-        )
-
-    idata = fit_laplace(
-        optimized_point, simp_model, transform_samples=transform_samples, progressbar=False
-    )
-
-    np.testing.assert_allclose(np.mean(idata.posterior.mu, axis=1), np.full((2,), 3), atol=0.1)
-    np.testing.assert_allclose(np.mean(idata.posterior.sigma, axis=1), np.full((2,), 1.5), atol=0.1)
